src/agents/aws_lambda/requirements.txt
# AWS Lambda dependencies for ResumeRank agents
# These are packaged into a Lambda Layer

# AWS SDK
boto3>=1.34.0

# PDF parsing
pdfplumber>=0.10.4

# DOCX parsing
python-docx>=1.1.0

# Data validation
pydantic>=2.5.0
